Refactoring Category name
