+++ client/app/components/transactions/transaction-list.tsx
@@ -1,6 +1,6 @@
 import { Transaction } from "@/app/models/transaction";
 import moment from "moment";
-import CategoryTag from "../categories/category";
+import CategoryTag from "../categories/CategoryTag";
 import classes from "./transaction-list.module.css";
 
 export default function TransactionList(props: {
